+++ Assets/Scripts/Other/DayController.cs
@@ -27,8 +27,8 @@ public class DayController : MonoBehaviour
 
     public void ChangeDay(int newState)
     {
-        anims.SetInteger("time", (int)newState);
-        dayState= (DayTime)newState;
+        anims.SetInteger("time", newState);
+        dayState = (DayTime)newState;
         counterTaskDay = 0;
         counter = newState;
         switch (dayState)
@@ -48,9 +48,9 @@ public class DayController : MonoBehaviour
         }
     }
 
-    IEnumerator Delay() 
+    IEnumerator Delay()
     {
-        yield return new WaitWhile(()=>GameManager.GetManager().dialogueManager.waitDialogue);
+        yield return new WaitWhile(() => GameManager.GetManager().dialogueManager.waitDialogue);
         GameManager.GetManager().dialogueManager.SetDialogue("PonerseATrabajar", delegate
         {
             GameManager.GetManager().blockController.UnlockAll(DayTime.Tarde);
@@ -85,7 +85,7 @@ public class DayController : MonoBehaviour
     {
         counterTaskDay++;
         if (dayState == DayTime.Noche)
-            GameManager.GetManager().dialogueManager.SetDialogue("Anochece", canRepeat:true);
+            GameManager.GetManager().dialogueManager.SetDialogue("Anochece", canRepeat: true);
 
         print(counterTaskDay + " nuevo stado" + (counterTaskDay % 5 == 0));
         if (counterTaskDay >= maxTasks)
